tests/wpt/tests/html/canvas/offscreen/text/2d.text.measure.width.basic.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<title>OffscreenCanvas test: 2d.text.measure.width.basic</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.text.measure.width.basic</h1>
<p class="desc">The width of character is same as font used</p>


<script>
promise_test(async t => {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  var f = new FontFace("CanvasTest", "url('/fonts/CanvasTest.ttf')");
  f.load();
  document.fonts.add(f);
  await document.fonts.ready;
  ctx.font = '50px CanvasTest';
  _assertSame(ctx.measureText('A').width, 50, "ctx.measureText('A').width", "50");
  _assertSame(ctx.measureText('AA').width, 100, "ctx.measureText('AA').width", "100");
  _assertSame(ctx.measureText('ABCD').width, 200, "ctx.measureText('ABCD').width", "200");

  ctx.font = '100px CanvasTest';
  _assertSame(ctx.measureText('A').width, 100, "ctx.measureText('A').width", "100");

}, "The width of character is same as font used");
</script>
